+++ Code/app/mods_tab.py
@@ -8,7 +8,7 @@ from Code.loc import Localization as loc
 from Code.package import ModUnit
 
 
-class ModTab:
+class ModsTab:
     dragged_mod_id = None
     active_mod_search_text = ""
     inactive_mod_search_text = ""
@@ -21,7 +21,7 @@ class ModTab:
             with dpg.group(horizontal=True):
                 dpg.add_button(
                     label=loc.get_string("btn-sort-mods"),
-                    callback=ModTab.sort_active_mods,
+                    callback=ModsTab.sort_active_mods,
                     tag="sort_button",
                 )
                 with dpg.tooltip("sort_button"):
@@ -76,12 +76,12 @@ class ModTab:
                     dpg.add_input_text(
                         tag="active_mod_search_tag",
                         hint=loc.get_string("input-hint-search"),
-                        callback=ModTab.on_search_changed,
+                        callback=ModsTab.on_search_changed,
                         user_data="active",
                     )
                     with dpg.child_window(
                         tag="active_mods_child",
-                        drop_callback=ModTab.on_mod_dropped,
+                        drop_callback=ModsTab.on_mod_dropped,
                         user_data="active",
                         payload_type="MOD_DRAG",
                     ):
@@ -92,43 +92,43 @@ class ModTab:
                     dpg.add_input_text(
                         tag="inactive_mod_search_tag",
                         hint=loc.get_string("input-hint-search"),
-                        callback=ModTab.on_search_changed,
+                        callback=ModsTab.on_search_changed,
                         user_data="inactive",
                     )
                     with dpg.child_window(
                         tag="inactive_mods_child",
-                        drop_callback=ModTab.on_mod_dropped,
+                        drop_callback=ModsTab.on_mod_dropped,
                         user_data="inactive",
                         payload_type="MOD_DRAG",
                     ):
                         pass
 
-        ModTab.render_mods()
+        ModsTab.render_mods()
 
     @staticmethod
     def on_search_changed(sender, app_data, user_data):
         if user_data == "active":
-            ModTab.active_mod_search_text = app_data.lower()
+            ModsTab.active_mod_search_text = app_data.lower()
 
         elif user_data == "inactive":
-            ModTab.inactive_mod_search_text = app_data.lower()
+            ModsTab.inactive_mod_search_text = app_data.lower()
 
-        ModTab.render_mods()
+        ModsTab.render_mods()
 
     @staticmethod
     def render_mods():
         ModManager.process_errors()
         dpg.delete_item("active_mods_child", children_only=True)
         for mod in ModManager.active_mods:
-            if ModTab.active_mod_search_text in mod.name.lower():
-                ModTab.add_movable_mod(mod, "active", "active_mods_child")
+            if ModsTab.active_mod_search_text in mod.name.lower():
+                ModsTab.add_movable_mod(mod, "active", "active_mods_child")
 
         dpg.delete_item("inactive_mods_child", children_only=True)
         for mod in ModManager.inactive_mods:
-            if ModTab.inactive_mod_search_text in mod.name.lower():
-                ModTab.add_movable_mod(mod, "inactive", "inactive_mods_child")
+            if ModsTab.inactive_mod_search_text in mod.name.lower():
+                ModsTab.add_movable_mod(mod, "inactive", "inactive_mods_child")
 
-        error_count, warning_count = ModTab.count_mods_with_issues()
+        error_count, warning_count = ModsTab.count_mods_with_issues()
         dpg.set_value(
             "error_count_text", loc.get_string("error-count", count=error_count)
         )
@@ -145,7 +145,7 @@ class ModTab:
             dpg.add_text(
                 mod.name,
                 tag=mod_name_tag,
-                drop_callback=ModTab.on_mod_dropped,
+                drop_callback=ModsTab.on_mod_dropped,
                 payload_type="MOD_DRAG",
                 user_data={"mod_id": mod.id, "status": status},
             )
@@ -214,7 +214,7 @@ class ModTab:
 
                 dpg.add_button(
                     label=loc.get_string("btn-show-full-details"),
-                    callback=lambda: ModTab.show_details_window(mod),
+                    callback=lambda: ModsTab.show_details_window(mod),
                 )
 
             with dpg.drag_payload(
@@ -366,12 +366,12 @@ class ModTab:
             else:
                 ModManager.move_inactive_mod_to_end(dragged_mod_id)
 
-        ModTab.render_mods()
+        ModsTab.render_mods()
 
     @staticmethod
     def sort_active_mods():
         ModManager.sort()
-        ModTab.render_mods()
+        ModsTab.render_mods()
 
     @staticmethod
     def count_mods_with_issues():
